Moved sampleIntents to a new file
